- flags: Update GBR to year 1801 version of the flag
The previous flag of Great Britain (also known as "King's Colours") was used
between years 1606 and 1801 and adopted as a national flag in 1707. Alas, the
Kingdom of Great Britain is historical and no more since 31 December 1800.

The current flag of the United Kingdom (also known as "Union Jack" or "Union
Flag") has been in use since year 1801. This version has Saint Patrick's red
saltire on it, which was notably missing from the previous flag Doomseeker used.

This replaces King's Colours (1606/1707) with Union Jack (1801).

Fixes: #3365
Reported-by: AOSP