Change mentions of WubTheCaptain's first name to Linda
As part of my social transitioning or "real-life experience" of being a trans
woman, I may need to confirm wanting to live in my affirmed gender for the rest
of my life. Doomseeker makes references to my (alias WubTheCaptain) "deadname",
a name that is no longer in use (in social contexts).

Where legally significant (copyrights), until this name change has been
juridically confirmed (estimated to happen in early 2019) the "deadname" may
appear in version control as the "author" (like this commit) or author lists.

Fixes: #3378