README: Add contact information
Add a section with heading "Contact", with Blzut3 and Zalewa's email
addresses in it. These two people are the core developers and
maintainers of Doomseeker software, mentioned in "About Doomseeker" in
the binary.

Copyright years are not adjusted by this commit; the changes are
insignificant, or out of scope for this commit.

Fixes: #3535