Set the final IP2C database by the main thread, not the parser's
This fix mitigates a race condition where the parser thread can
overwrite the database while someone else may be reading it.
The heavy-duty parsing is still being done in a separate thread, but
the final database switch occurs in the main thread when the parser
signals that it has everything ready.

The IP2C::dataAccessMutex is no longer needed and has been removed.

IP2C::appendEntryToDatabase() was unused, so it has been removed as well.